refactor(shared-kernel): player-type calculator moves in from the recap slice
RecapPlayerType and RecapPlayerTypeCalculator relocate verbatim from
PlayerProgress (Contracts.Recap / Domain.Recap, internal) to
SharedKernel.Enums / SharedKernel.Models, public. The classifier is a pure
function over PhoenixScore collections and the Official Mirror overhaul
labels leaderboard players with it next; PlayerRecap and the recap saga
keep compiling through their existing SharedKernel usings.
